requirements.txt
Flask
pillow
qrcode
gspread
google-auth
google-api-python-client
